+++ packages/rotate/src/types/StoreProps.ts
@@ -10,6 +10,6 @@ import { RotateDirection } from '@react-pdf-viewer/core';
 
 export interface StoreProps {
     rotation: number;
-    rotate?(degrees: number): void;
+    rotate?(direction: RotateDirection): void;
     rotatePage?(pageIndex: number, direction: RotateDirection): void;
 }